Rename license file and update composer.json.
Although both are correct spellings (US English versus British English),
GitHub and Packagist look for a "LICENSE" file (US English), rather than a
"LICENCE" file (British English), for determining which licensing
information to display for a repository.

Packagist also prefers that the type of licensing be stated explicitly in
the composer.json file.